+++ src/main/java/com/gmail/val59000mc/scoreboard/ScoreboardManager.java
@@ -72,7 +72,7 @@ public class ScoreboardManager {
 		}
 
 		if (returnString.contains("%teamColor%")){
-			returnString = returnString.replace("%teamColor%",uhcPlayer.getTeam().getPrefix());
+			returnString = returnString.replace("%teamColor%", uhcPlayer.getTeam().getColor() + "\u25A0");
 		}
 
 		if (returnString.contains("%border%")){